requirements.txt
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
Pillow==10.2.0
pypandoc==1.12
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
ffmpeg-python==0.2.0
py7zr==0.21.0
rarfile==4.1
requests==2.31.0
psutil==5.9.8
packaging==23.2